% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-additional-functions.R
\name{assign.by.size}
\alias{assign.by.size}
\title{Assign categories to features based on size & shape}
\usage{
assign.by.size(site, lower = 3, mid = 50, upper = 100, ratio = 0.8,
  density = 0.2, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{lower}{Cell count threshold below which a feature is classed as noise.}

\item{mid}{Cell count threshold below which a feature is classed as a post-hole.}

\item{upper}{Cell count threshold above which a feature is classed as a linear feature of the map.}

\item{ratio}{Ratio (of longest to shortest edge) above which}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types. Also plots the new feature types.
}
\description{
For each feature, get the ratio of the longest to shortest bounding edge. Small features with a ratio close to 1 are set as post-holes; slightly larger as annotations; the largest as linear features.
}
\examples{
assign.by.size(genlis)
}

