% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R, R/get-features.R
\name{extend.annotations}
\alias{extend.annotations}
\title{Identify any features that lie directly between two annotations}
\usage{
extend.annotations(site, plot = T)

extend.annotations(site, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}

\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.

New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.
}
\description{
For each feature classed as an annotation, the function extends the feature's footprint directly left, right, up and down. Any features that appear twice in these extensions (ie. features that are directly horizontally or vertically between two annotation features) are also classed as annotations. Helps with picking up broken line boundaries, dots in text etc.

For each feature classed as an annotation, the function extends the feature's footprint directly left, right, up and down. Any features that appear twice in these extensions (ie. features that are directly horizontally or vertically between two annotation features) are also classed as annotations. Helps with picking up broken line boundaries, dots in text etc.
}
\examples{
extend.annotations(genlis)
extend.annotations(genlis)
}

