% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{feature.closing}
\alias{feature.closing}
\title{Find features whose shape differs after a morphological closing}
\usage{
feature.closing(site, progress.bar = F, plot.progress = F, rm = 1)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{progress.bar}{Boolean: display progress bar or not? Default is F.}

\item{plot.progress}{Boolean: plot features and colour those that have been checked, or not? Default is F}

\item{rm}{Radius modifier: how many pixels' widths should be used as a radius for the dilation and erosion processes of the closing? Default is 1. Higher values will lead to a greater degree of smoothing.}
}
\value{
Matrix showing the cluster ID and area before and after the closing operation, and a list of polygon objects representing the features after closing.
}
\description{
For all unclassified features, removes holes from polygons and performs a morphological closing with radius 1 pixel, and determines which features are unaffected by this process.
}
\examples{
close.features(sparse.shapes.classified, plot.progress = T)
}

