% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{fill.broken.boundary}
\alias{fill.broken.boundary}
\title{Fill in broken (-.-) line boundary}
\usage{
fill.broken.boundary(site, plot.progress = F, s = 0.2)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot.progress}{Boolean indicator: should the function's progress be displayed on a plot?}

\item{s}{Threshold value between 0 and 1, indicating how 'sparse' a feature must be to be considered as a line. Default is 0.2.}
}
\value{
Vector of ids of features belonging to the boundary.
}
\description{
Find all linear features in site and identify any features lying along their longest axis. Any features aligned to 2 or more linear features are considered to be part of a broken-line boundary.
}
\examples{
# find boundary features
fill.broken.boundary(genlis, plot.progress = T)
}

