% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{filter.by.distance}
\alias{filter.by.distance}
\title{Identify remote points}
\usage{
filter.by.distance(pts, plot = F)
}
\arguments{
\item{pts}{A two-column matrix containing the coordinates of the points to be compared.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}
}
\value{
A Boolean array of length n that can be used to filter points or angles. TRUE indicates that the point is particularly remote from its nearest neighbours; FALSE indicates that point is relatively close to its neighbours.
}
\description{
For a set of x and y coordinates, find the distance to the nearest neighbour. Extremely isolated points (outliers) are defined in the same way as in a boxplot: any points whose nearest-neighbour distance lies more than 1.5 x the IQR above the third quartile.
}
\examples{
nn.filter <- filter.by.distance(centres, plot = T)
}

