% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features.R
\name{get.horizontals}
\alias{get.horizontals}
\title{Identify horizontal features}
\usage{
get.horizontals(site, l = 7, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{l}{Length of black pixels to be tested for (should be an odd number, but function will add 1 if an even number is supplied). Default is 9.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.
}
\description{
Identify unclassified/post-hole features that contain a horizontal strip of black pixels, and set as annotations.
}
\examples{
get.horizontal(genlis)
}

