% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{get.orientation}
\alias{get.orientation}
\title{Identify N-S marker on map}
\usage{
get.orientation(site)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}
}
\value{
The angle (in radians) of the N-S axis.
}
\description{
Identify the N-S marker on the image and calculate the angle of the N-S axis. Since the marker shape varies from plan to plan, it must be selected manually.
}
\examples{
get.orientation(genlis)
}

