% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features.R
\name{get.postholes}
\alias{get.postholes}
\title{Extract centre-points of post-holes}
\usage{
get.postholes(site, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}
}
\value{
Matrix of coordinates showing the midpoints of the features selected
}
\description{
For each post-hole feature identified, get the x and y coordinates of the centre of the feature.
}
\examples{
get.postholes(genlis)
}

