% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/save-and-retrieve.R
\name{import.map}
\alias{import.map}
\title{Import map from jpeg}
\usage{
import.map(jpg.file, plot = T, threshold = 0.2)
}
\arguments{
\item{jpg.file}{String containing the full name (including path, if not in current working directory) of the jpeg to import.}

\item{plot}{Logical indicating whether to draw a plot of the imported data or not. Defaults to T.}

\item{threshold}{Numeric between 0 and 1, indicating the threshold at which points are divided into white or black. A higher threshold will only convert darker objects to black pixels, while a lower threshold will convert fainter objects to black. defaults to 0.2.}
}
\value{
A new feature set: a list containing a raster of numbered features, and a matrix assigning a type to each feature.
}
\description{
Load a jpeg map image and convert to a black-and-white binary raster
}
\examples{
genlis <- import.map("Genlis-cropped.jpg", plot = T, threshold = 0.2)
}

