% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/save-and-retrieve.R
\name{load.features}
\alias{load.features}
\title{Load a feature set}
\usage{
load.features(name)
}
\arguments{
\item{name}{Name of files that was to save the feature set. Should not include file extension, since it will be used to refer to multiple files}
}
\value{
A new feature set: a list containing a raster of numbered features, and a matrix assigning a type to each feature.
}
\description{
Load previously-saved components of a feature set and assign to an object.
}
\examples{
genlis <- load.features("Genlis")
}

