% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-additional-functions.R
\name{manually.set.feature.type}
\alias{manually.set.feature.type}
\title{Manually reclassify a single feature}
\usage{
manually.set.feature.type(site, replot = F)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{replot}{Boolean: do you want to redraw the site before selecting the feature to reclassify?}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.
}
\description{
A last resort: manually select a single feature from the map and reclassify it as a post-hole, an annotation, or a linear feature of the map.
}
\examples{
manually.set.feature.type(genlis)
}

