% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R, R/get-features.R
\name{overlay.postholes}
\alias{overlay.postholes}
\title{Overlay map showing postholes}
\usage{
overlay.postholes(site, postholes)

overlay.postholes(site, postholes)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{postholes}{Matrix of coordinates showing the midpoints of the posthole features to be plotted.}

\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{postholes}{Matrix of coordinates showing the midpoints of the posthole features to be plotted.}
}
\description{
Plot the site's features, with the midpoints of the identified posthole features overlaid.

Plot the site's features, with the midpoints of the identified posthole features overlaid.
}
\examples{
get.postholes(genlis)
get.postholes(genlis)
}

