% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{remove.annotations}
\alias{remove.annotations}
\title{Remove annotations}
\usage{
remove.annotations(site, plot = T, show.progress = T, remove.similar = F)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}

\item{remove.similar}{Boolean indicator: If T, after finding all sequences of horizontal features, will also add in any features that are within a 1-cell tolerance of the modal height and width of the features already identified.}

\item{show.process}{Boolean indicator: show progress bar or not? Defaults to T.}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.
}
\description{
Identify horizontal sequences of similarly-sized features as annotations.
}
\examples{
remove.annotations(genlis)
}

