% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features.R
\name{remove.isolated}
\alias{remove.isolated}
\title{Remove isolated features}
\usage{
remove.isolated(site, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}
}
\description{
For each post-hole feature identified, find the distance to the nearest neighbour. Extremely isolated points (outliers) are defined in the same way as in a boxplot: any points whose nearest-neighbour distance lies more than 1.5 x the IQR above the third quartile.
}
\examples{
remove.isolated(genlis)
}

