% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-key-functions.R
\name{remove.tall.features}
\alias{remove.tall.features}
\title{Remove unusually tall features}
\usage{
remove.tall.features(site, plot = T)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{plot}{Boolean indicator: should the new set of features be displayed on a plot?}
}
\value{
New feature set list containing the original raster of numbered features and a matrix of newly assigned feature types.
}
\description{
Calculated the height of all unclassified/post-hole features and finds the threshold for an outlier in the same way as for a boxplot (1.5 x IQR above the upper quartile of the data); any features taller than this are set as annotations.
}
\examples{
remove.tall.features(genlis)
}

