% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/save-and-retrieve.R
\name{save.features}
\alias{save.features}
\title{Save a feature set}
\usage{
save.features(obj.to.save, name)
}
\arguments{
\item{obj.to.save}{Feature set to be saved}

\item{name}{Name of files to be used to save the feature set. Should not include file extension, since it will be used to save multiple files}
}
\description{
Save the components of a feature set for later retrieval.
}
\examples{
save.features(genlis, "Genlis")
}

