% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get-features-additional-functions.R
\name{select.feature}
\alias{select.feature}
\title{Identify a feature from the map}
\usage{
select.feature(site, replot = F)
}
\arguments{
\item{site}{Feature set list, containing a raster of all features, and a matrix assigning each feature to a particular type.}

\item{replot}{Boolean indicator: should the map be redrawn, or can the function use the currently drawn plot?}
}
\description{
Get the id number and x, y coordinates from a particular feature by clicking on it on the map. Can be used in conjunction with \code{feature.dims} to check the height, density and other characteristics of a particular feature.
}
\examples{
select.feature(genlis, replot = F)
}

