% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/save-and-retrieve.R
\name{show.features}
\alias{show.features}
\title{Plot features & feature types}
\usage{
show.features(feature.set)
}
\arguments{
\item{feature.set}{The feature set to be plotted}
}
\description{
Plot all features in set, coloured according to their assigned type.
}
\examples{
show.features(genlis)
}

